requirements.txt
# aespa Brand Intelligence System - Dependencies

# Search & Scraping
google-search-results>=2.4.2    # SerpAPI SDK
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tavily-python>=0.3.3
apify-client>=1.6.0

# LLM
google-genai>=1.0.0             # Gemini API SDK (新版)

# Report Generation
Jinja2>=3.1.0

# Environment & Config
python-dotenv>=1.0.0
